src/__tests__/interfaces/interfaces.test.ts
import { PutItemOutput } from 'aws-sdk/clients/dynamodb';
import {
	RegExValidators,
	SerialNumberMappings,
	NoteDetails,
	MatchedDetail,
	DenominationDetail,
	DetectedText,
	ExtendedDetectedText,
	UploadData,
	FederalReserveMapping,
	FileChecker,
	FileOperations,
	AWSService,
	MockedDynamoDbResponse,
	DynamoDbResponse,
	RouteError,
} from '../../interfaces/interfaces';

/**
 * Mock the AWS SDK for testing purposes
 */
jest.mock('aws-sdk', () => ({
	S3: jest.fn().mockImplementation(() => ({
		putObject: jest.fn().mockReturnThis(),
		promise: jest.fn().mockResolvedValue({ Location: 's3Url' }), // Return a dummy S3 URL
	})),
}));

describe('Interfaces', () => {
	/**
	 * Test suite for the SerialNumberMappings interface
	 */
	describe('SerialNumberMappings', () => {
		/**
		 * Test case to ensure the SerialNumberMappings interface has string keys and an array of objects with specific properties
		 */
		it('should have string keys and an array of objects with specific properties', () => {
			const serialNumberMappings: SerialNumberMappings = {
				key1: [{ pattern: /pattern1/, denomination: 'denom1', seriesYear: 'year1', treasurer: 'treasurer1', secretary: 'secretary1' }],
				key2: [{ pattern: /pattern2/, denomination: 'denom2', seriesYear: 'year2', treasurer: 'treasurer2', secretary: 'secretary2' }],
			};
			expect(serialNumberMappings).toHaveProperty('key1');
			expect(serialNumberMappings).toHaveProperty('key2');
			expect(serialNumberMappings.key1).toBeInstanceOf(Array);
			expect(serialNumberMappings.key2).toBeInstanceOf(Array);
			expect(serialNumberMappings.key1[0]).toHaveProperty('pattern');
			expect(serialNumberMappings.key1[0]).toHaveProperty('denomination');
			expect(serialNumberMappings.key1[0]).toHaveProperty('seriesYear');
			expect(serialNumberMappings.key1[0]).toHaveProperty('treasurer');
			expect(serialNumberMappings.key1[0]).toHaveProperty('secretary');
			expect(serialNumberMappings.key1[0].pattern).toBeInstanceOf(RegExp);
		});
	});

	/**
	 * Test suite for the FederalReserveMapping interface
	 */
	describe('FederalReserveMapping', () => {
		/**
		 * Test case to ensure the FederalReserveMapping interface has string keys and string values
		 */
		it('should have string keys and string values', () => {
			const federalReserveMapping: FederalReserveMapping = {
				key1: 'value1',
				key2: 'value2',
			};
			expect(federalReserveMapping).toHaveProperty('key1');
			expect(federalReserveMapping).toHaveProperty('key2');
			expect(typeof federalReserveMapping.key1).toBe('string');
			expect(typeof federalReserveMapping.key2).toBe('string');
		});
	});

	/**
	 * Test suite for the NoteDetails interface
	 */
	describe('NoteDetails', () => {
		/**
		 * Test case to ensure the NoteDetails interface has specific properties
		 */
		it('should have specific properties', () => {
			const noteDetails: NoteDetails = {
				validDenomination: 'denom',
				frontPlateId: 'frontId',
				SerialPatternMatch: 'pattern',
				validSerialNumberPattern: 'pattern',
				federalReserveId: 'id',
				notePositionId: 'position',
				seriesYear: 'year',
				treasurer: 'treasurer',
				secretary: 'secretary',
			};
			expect(noteDetails).toHaveProperty('validDenomination');
			expect(noteDetails).toHaveProperty('frontPlateId');
			expect(noteDetails).toHaveProperty('SerialPatternMatch');
			expect(noteDetails).toHaveProperty('validSerialNumberPattern');
			expect(noteDetails).toHaveProperty('federalReserveId');
			expect(noteDetails).toHaveProperty('notePositionId');
			expect(noteDetails).toHaveProperty('seriesYear');
			expect(noteDetails).toHaveProperty('treasurer');
			expect(noteDetails).toHaveProperty('secretary');
		});
	});

	/**
	 * Test suite for the RegExValidators interface
	 */
	describe('RegExValidators', () => {
		/**
		 * Test case to ensure the RegExValidators interface has string keys and RegExp values
		 */
		it('should have string keys and RegExp values', () => {
			const regexValidators: RegExValidators = {
				key1: /pattern1/,
				key2: /pattern2/,
			};
			expect(regexValidators).toHaveProperty('key1');
			expect(regexValidators).toHaveProperty('key2');
			expect(regexValidators.key1).toBeInstanceOf(RegExp);
			expect(regexValidators.key2).toBeInstanceOf(RegExp);
		});
	});

	/**
	 * Test suite for the MatchedDetail interface
	 */
	describe('MatchedDetail', () => {
		/**
		 * Test case to ensure the MatchedDetail interface has specific properties
		 */
		it('should have specific properties', () => {
			const matchedDetail: MatchedDetail = {
				seriesYear: 'year',
				treasurer: 'treasurer',
				secretary: 'secretary',
			};
			expect(matchedDetail).toHaveProperty('seriesYear');
			expect(matchedDetail).toHaveProperty('treasurer');
			expect(matchedDetail).toHaveProperty('secretary');
			expect(typeof matchedDetail.seriesYear).toBe('string');
			expect(typeof matchedDetail.treasurer).toBe('string');
			expect(typeof matchedDetail.secretary).toBe('string');
		});
	});

	/**
	 * Test suite for the DenominationDetail interface
	 */
	describe('DenominationDetail', () => {
		/**
		 * Test case to ensure the DenominationDetail interface has specific properties
		 */
		it('should have specific properties', () => {
			const denominationDetail: DenominationDetail = {
				seriesYear: 'year',
				treasurer: 'treasurer',
				secretary: 'secretary',
				pattern: /regexPattern/,
			};
			expect(denominationDetail).toHaveProperty('seriesYear');
			expect(denominationDetail).toHaveProperty('treasurer');
			expect(denominationDetail).toHaveProperty('secretary');
			expect(denominationDetail).toHaveProperty('pattern');
			expect(typeof denominationDetail.seriesYear).toBe('string');
			expect(typeof denominationDetail.treasurer).toBe('string');
			expect(typeof denominationDetail.secretary).toBe('string');
			expect(denominationDetail.pattern).toBeInstanceOf(RegExp);
		});
	});

	/**
	 * Test suite for the DetectedText interface
	 */
	describe('DetectedText', () => {
		/**
		 * Test case to ensure the DetectedText interface has specific properties
		 */
		it('should have specific properties', () => {
			const detectedText: DetectedText = {
				validDenomination: '20',
				frontPlateId: 'frontPlateId',
				SerialPatternMatch: 'SerialPatternMatch',
				serialNumber: 'serialNumber',
				federalReserveId: 'federalReserveId',
				notePositionId: 'notePositionId',
			};
			expect(detectedText).toHaveProperty('validDenomination');
			expect(detectedText).toHaveProperty('frontPlateId');
			expect(detectedText).toHaveProperty('SerialPatternMatch');
			expect(detectedText).toHaveProperty('serialNumber');
			expect(detectedText).toHaveProperty('federalReserveId');
			expect(detectedText).toHaveProperty('notePositionId');
			expect(typeof detectedText.validDenomination).toBe('string');
			expect(typeof detectedText.frontPlateId).toBe('string');
			expect(typeof detectedText.SerialPatternMatch).toBe('string');
			expect(typeof detectedText.serialNumber).toBe('string');
			expect(typeof detectedText.federalReserveId).toBe('string');
			expect(typeof detectedText.notePositionId).toBe('string');
		});
	});

	/**
	 * Test suite for the ExtendedDetectedText interface
	 */
	describe('ExtendedDetectedText', () => {
		/**
		 * Test case to ensure the ExtendedDetectedText interface extends DetectedText and MappedData interfaces
		 */
		it('should extend DetectedText and MappedData interfaces', () => {
			const extendedDetectedText: ExtendedDetectedText = {
				validDenomination: '20',
				frontPlateId: 'frontPlateId',
				SerialPatternMatch: 'SerialPatternMatch',
				serialNumber: 'serialNumber',
				federalReserveId: 'federalReserveId',
				notePositionId: 'notePositionId',
				seriesYear: 'year',
				treasurer: 'treasurer',
				secretary: 'secretary',
				federalReserveLocation: 'federalReserveLocation',
			};
			expect(extendedDetectedText).toHaveProperty('validDenomination');
			expect(extendedDetectedText).toHaveProperty('frontPlateId');
			expect(extendedDetectedText).toHaveProperty('SerialPatternMatch');
			expect(extendedDetectedText).toHaveProperty('serialNumber');
			expect(extendedDetectedText).toHaveProperty('federalReserveId');
			expect(extendedDetectedText).toHaveProperty('notePositionId');
			expect(extendedDetectedText).toHaveProperty('seriesYear');
			expect(extendedDetectedText).toHaveProperty('treasurer');
			expect(extendedDetectedText).toHaveProperty('secretary');
			expect(extendedDetectedText).toHaveProperty('federalReserveLocation');
		});
	});

	/**
	 * Test suite for the UploadData interface
	 */
	describe('UploadData', () => {
		/**
		 * Test case to ensure the UploadData interface has specific properties
		 */
		it('should have specific properties', () => {
			const uploadData: UploadData = {
				validDenomination: '20',
				frontPlateId: 'frontPlateId',
				SerialPatternMatch: 'SerialPatternMatch',
				serialNumber: 'serialNumber',
				federalReserveId: 'federalReserveId',
				notePositionId: 'notePositionId',
				seriesYear: 'year',
				treasurer: 'treasurer',
				secretary: 'secretary',
				federalReserveLocation: 'federalReserveLocation',
				s3Url: 's3Url',
			};
			expect(uploadData).toHaveProperty('validDenomination');
			expect(uploadData).toHaveProperty('frontPlateId');
			expect(uploadData).toHaveProperty('SerialPatternMatch');
			expect(uploadData).toHaveProperty('serialNumber');
			expect(uploadData).toHaveProperty('federalReserveId');
			expect(uploadData).toHaveProperty('notePositionId');
			expect(uploadData).toHaveProperty('seriesYear');
			expect(uploadData).toHaveProperty('treasurer');
			expect(uploadData).toHaveProperty('secretary');
			expect(uploadData).toHaveProperty('federalReserveLocation');
			expect(uploadData).toHaveProperty('s3Url');
		});
	});

	/**
	 * Test suite for the FileChecker interface
	 */
	describe('FileChecker', () => {
		/**
		 * Test case to ensure the FileChecker interface has a checkFileExists method
		 */
		it('should have a checkFileExists method', () => {
			const fileChecker: FileChecker = {
				checkFileExists: () => {
					// Omit the filePath parameter
					// Dummy implementation for test
					return true;
				},
			};
			expect(fileChecker.checkFileExists('/path/to/file')).toBe(true);
		});
	});

	describe('Unused Interfaces Tests', () => {
		/**
		 * Test suite for the FileOperations interface
		 */
		describe('FileOperations', () => {
			/**
			 * Test case to ensure the FileOperations interface has methods for checking file existence and reading files
			 */
			it('should have methods for checking file existence and reading files', () => {
				const fileOperations: FileOperations = {
					checkFileExists: () => {
						// Dummy implementation for test
						return true;
					},
					readFile: function (): string {
						throw new Error('Function not implemented.');
					},
				};
				// Call the method and assert its behavior
				expect(fileOperations.checkFileExists('/path/to/file')).toBe(true);
			});
		});

		/**
		 * Test suite for the AWSService interface
		 */
		describe('AWSService', () => {
			/**
			 * Test case to ensure the AWSService interface has methods for inserting into DynamoDB and saving to S3
			 */
			it('should have methods for inserting into DynamoDB and saving to S3', async () => {
				const awsService: AWSService = {
					insertIntoDynamo: async (): Promise<PutItemOutput> => {
						// Dummy implementation for test
						return {
							// Populate with appropriate properties of PutItemOutput
							// This is just an example, replace it with actual properties
							ConsumedCapacity: {
								CapacityUnits: 1,
							},
							ItemCollectionMetrics: {},
						};
					},

					saveToS3: async () => {
						// Dummy implementation for test
						return 's3Url';
					},
				};
				// Call the methods and assert their behavior
				const dataToInsert: UploadData = {
					validDenomination: '20',
					frontPlateId: 'frontPlateId',
					SerialPatternMatch: 'SerialPatternMatch',
					serialNumber: 'serialNumber',
					federalReserveId: 'federalReserveId',
					notePositionId: 'notePositionId',
					seriesYear: 'year',
					treasurer: 'treasurer',
					secretary: 'secretary',
					federalReserveLocation: 'federalReserveLocation',
					s3Url: 's3Url',
				};
				// Call the methods and assert their behavior
				const dynamoDbResponse = await awsService.insertIntoDynamo(dataToInsert);
				const s3Url = await awsService.saveToS3('filePath', 'Key');

				// Adjust assertions based on the actual structure of the response
				expect(dynamoDbResponse).toBeDefined(); // Check if the response exists
				expect(s3Url).toBe('s3Url');
			});
		});

		/**
		 * Test suite for the RouteError interface
		 */
		describe('RouteError', () => {
			/**
			 * Test case to ensure the RouteError interface is an instance of Error and has a status property
			 */
			it('should be an instance of Error and have a status property', () => {
				// Create an instance of the custom error class RouteError
				const routeError: RouteError = new RouteError('Route error', 404);
				// Assert the properties of the error
				expect(routeError).toBeInstanceOf(Error);
				expect(routeError).toHaveProperty('status', 404);
			});
		});

		/**
		 * Test suite for the MockedDynamoDbResponse interface
		 */
		describe('MockedDynamoDbResponse', () => {
			/**
			 * Test case to ensure the MockedDynamoDbResponse interface has specific properties and a method item
			 */
			it('should have specific properties and a method item', () => {
				const mockedDynamoDbResponse: MockedDynamoDbResponse = {
					status: 'success',
					item: (item: UploadData) => ({ status: 'success', item }),
					s3Url: '',
					validDenomination: '',
					frontPlateId: '',
					SerialPatternMatch: '',
					serialNumber: '',
					federalReserveId: '',
					notePositionId: '',
					seriesYear: '',
					treasurer: '',
					secretary: '',
					federalReserveLocation: '',
				};
				// Assert the properties and method of the mocked response
				expect(mockedDynamoDbResponse).toHaveProperty('status', 'success');
				expect(typeof mockedDynamoDbResponse.item).toBe('function');
			});
		});

		/**
		 * Test suite for the DynamoDbResponse interface
		 */
		describe('DynamoDbResponse', () => {
			/**
			 * Test case to ensure the DynamoDbResponse interface has specific properties
			 */
			it('should have specific properties', () => {
				const dynamoDbResponse: DynamoDbResponse = {
					status: 'success',
					item: {
						validDenomination: '20',
						frontPlateId: 'frontPlateId',
						SerialPatternMatch: 'SerialPatternMatch',
						serialNumber: 'serialNumber',
						federalReserveId: 'federalReserveId',
						notePositionId: 'notePositionId',
						seriesYear: 'year',
						treasurer: 'treasurer',
						secretary: 'secretary',
						federalReserveLocation: 'federalReserveLocation',
						s3Url: 's3Url',
					},
				};
				expect(dynamoDbResponse).toHaveProperty('status', 'success');
				expect(dynamoDbResponse.item).toEqual({
					validDenomination: '20',
					frontPlateId: 'frontPlateId',
					SerialPatternMatch: 'SerialPatternMatch',
					serialNumber: 'serialNumber',
					federalReserveId: 'federalReserveId',
					notePositionId: 'notePositionId',
					seriesYear: 'year',
					treasurer: 'treasurer',
					secretary: 'secretary',
					federalReserveLocation: 'federalReserveLocation',
					s3Url: 's3Url',
				});
			});
		});
	});
});
